+++ src/test/java/org/opensearch/neuralsearch/query/NeuralQueryBuilderTests.java
@@ -646,10 +646,10 @@ public class NeuralQueryBuilderTests extends OpenSearchTestCase {
     @SneakyThrows
     public void testRewrite_whenVectorSupplierNull_thenSetVectorSupplier() {
         NeuralQueryBuilder neuralQueryBuilder = new NeuralQueryBuilder().fieldName(FIELD_NAME).queryText(QUERY_TEXT).modelId(MODEL_ID).k(K);
-        List<Float> expectedVector = Arrays.asList(1.0f, 2.0f, 3.0f, 4.0f, 5.0f);
+        List<Number> expectedVector = Arrays.asList(1.0f, 2.0f, 3.0f, 4.0f, 5.0f);
         MLCommonsClientAccessor mlCommonsClientAccessor = mock(MLCommonsClientAccessor.class);
         doAnswer(invocation -> {
-            ActionListener<List<Float>> listener = invocation.getArgument(2);
+            ActionListener<List<Number>> listener = invocation.getArgument(2);
             listener.onResponse(expectedVector);
             return null;
         }).when(mlCommonsClientAccessor).inferenceSentences(any(), anyMap(), any());
@@ -682,10 +682,10 @@ public class NeuralQueryBuilderTests extends OpenSearchTestCase {
             .queryImage(IMAGE_TEXT)
             .modelId(MODEL_ID)
             .k(K);
-        List<Float> expectedVector = Arrays.asList(1.0f, 2.0f, 3.0f, 4.0f, 5.0f);
+        List<Number> expectedVector = Arrays.asList(1.0f, 2.0f, 3.0f, 4.0f, 5.0f);
         MLCommonsClientAccessor mlCommonsClientAccessor = mock(MLCommonsClientAccessor.class);
         doAnswer(invocation -> {
-            ActionListener<List<Float>> listener = invocation.getArgument(2);
+            ActionListener<List<Number>> listener = invocation.getArgument(2);
             listener.onResponse(expectedVector);
             return null;
         }).when(mlCommonsClientAccessor).inferenceSentences(any(), anyMap(), any());